feat: start of testing
