+++ src/AntTargetRunner.js
@@ -1,6 +1,6 @@
 const vscode = require('vscode')
 const dotenv = require('dotenv')
-const filehelper = require('./filehelper')
+const fileHelper = require('./fileHelper')
 const fs = require('fs')
 const path = require('path')
 
@@ -48,17 +48,17 @@ module.exports = class AntTargetRunner {
       } else {
         this.antExecutable = 'ant.sh'
       }
-      // if (!filehelper.pathExists(this.antExecutable) && filehelper.pathExists(this.antHome)) {
+      // if (!fileHelper.pathExists(this.antExecutable) && fileHelper.pathExists(this.antHome)) {
       //   const joinedPath = path.join(this.antHome, 'bin', this.antExecutable)
-      //   if (filehelper.pathExists(joinedPath)) {
+      //   if (fileHelper.pathExists(joinedPath)) {
       //     this.antExecutable = '"' + joinedPath + '"'
       //   }
       // }
     }
 
     // enable WinColorLogger?
-    if (process.platform === 'win32' && filehelper.pathExists(this.antHome)) {
-      this.useWinColorLogger = filehelper.pathExists(path.join(this.antHome, 'lib', 'WinColorLogger.jar'))
+    if (process.platform === 'win32' && fileHelper.pathExists(this.antHome)) {
+      this.useWinColorLogger = fileHelper.pathExists(path.join(this.antHome, 'lib', 'WinColorLogger.jar'))
     }
 
     this.buildFileDirectories = configOptions.get('buildFileDirectories', '.')
@@ -70,7 +70,7 @@ module.exports = class AntTargetRunner {
       this.envVarsFile = 'build.env'
     }
 
-    this.envVarsFile = await filehelper.findFirstFile(this.rootPath, this.buildFileDirectories.split(','), this.envVarsFile.split(','))
+    this.envVarsFile = await fileHelper.findFirstFile(this.rootPath, this.buildFileDirectories.split(','), this.envVarsFile.split(','))
 
     if (this.antTerminal) {
       this.antTerminal.dispose()
@@ -101,8 +101,8 @@ module.exports = class AntTargetRunner {
 
     if (!this.antTerminal) {
       var envVars = {}
-      if (this.envVarsFile && filehelper.pathExists(filehelper.getRootFile(this.rootPath, this.envVarsFile))) {
-        envVars = dotenv.parse(fs.readFileSync(filehelper.getRootFile(this.rootPath, this.envVarsFile)))
+      if (this.envVarsFile && fileHelper.pathExists(fileHelper.getRootFile(this.rootPath, this.envVarsFile))) {
+        envVars = dotenv.parse(fs.readFileSync(fileHelper.getRootFile(this.rootPath, this.envVarsFile)))
       }
 
       if (this.antHome) {
@@ -110,9 +110,9 @@ module.exports = class AntTargetRunner {
       }
 
       // add ant to path?
-      if (!filehelper.pathExists(this.antExecutable) && filehelper.pathExists(this.antHome)) {
+      if (!fileHelper.pathExists(this.antExecutable) && fileHelper.pathExists(this.antHome)) {
         const joinedPath = path.join(this.antHome, 'bin')
-        if (filehelper.pathExists(joinedPath)) {
+        if (fileHelper.pathExists(joinedPath)) {
           if (process.platform === 'win32') {
             envVars.Path = process.env.Path + ';' + joinedPath
           } else {
@@ -122,7 +122,7 @@ module.exports = class AntTargetRunner {
       }
 
       // use ansicon on win32?
-      if (process.platform === 'win32' && this.ansiconExe && filehelper.pathExists(this.ansiconExe)) {
+      if (process.platform === 'win32' && this.ansiconExe && fileHelper.pathExists(this.ansiconExe)) {
         if (envVars.ANT_ARGS === undefined) {
           envVars.ANT_ARGS = ' -logger org.apache.tools.ant.listener.AnsiColorLogger'
         }
@@ -160,7 +160,7 @@ module.exports = class AntTargetRunner {
   }
 
   revealDefinition (target) {
-    vscode.workspace.openTextDocument(filehelper.getRootFile(this.rootPath, target.sourceFile))
+    vscode.workspace.openTextDocument(fileHelper.getRootFile(this.rootPath, target.sourceFile))
       .then((document) => {
         return vscode.window.showTextDocument(document)
       })
