+++ src/BuildFileParser.js
@@ -1,20 +1,30 @@
-const filehelper = require('./filehelper')
+const fileHelper = require('./fileHelper')
 const fs = require('fs')
 const path = require('path')
 const xml2js = require('xml2js')
+const AntWrapper = require('./AntWrapper')
+const vscode = require('vscode')
 
 module.exports = class BuildFileParser {
-  constructor (rootPath) {
+  constructor (context, rootPath) {
     this.rootPath = rootPath
     this._parser = new xml2js.Parser()
     console.debug(this.rootPath)
     this.useAntToParse = true
+    this.antWrapper = new AntWrapper(context)
+    this.getConfigOptions()
+  }
+
+  getConfigOptions () {
+    let configOptions = vscode.workspace.getConfiguration('ant', null)
+
+    this.useAntForParsing = configOptions.get('parseBuildFileWithAnt', 'ant')
   }
 
   findBuildFile (searchDirectories, searchFileNames) {
     return new Promise(async (resolve, reject) => {
       try {
-        var filename = await filehelper.findFirstFile(this.rootPath, searchDirectories, searchFileNames)
+        var filename = await fileHelper.findFirstFile(this.rootPath, searchDirectories, searchFileNames)
         resolve(filename)
       } catch (error) {
         return reject(new Error('No build file found!'))
@@ -103,150 +113,28 @@ module.exports = class BuildFileParser {
     })
   }
 
+  async parseBuildFile (buildFileName) {
+    if (this.useAntForParsing) {
+      await this.parseBuildFileWithAnt(buildFileName)
+    } else {
+      await this.parseBuildFileDirect(buildFileName)
+    }
+  }
+
   async parseBuildFileWithAnt (buildFileName) {
     try {
-      const data = await this.spawnAnt(buildFileName)
-      const buildFile = await this.parseAntData(data)
+      const data = await this.antWrapper.spawnAnt(buildFileName)
+      const buildFile = this.antWrapper.parseAntData(data)
     } catch(err) {
       console.error(err)
     }
-    // data=> {parseData(data)},
-    // err=>  {console.error("async error:\n" + err)}
-    // )
   }
 
-  async parseAntData (data) {
-    const buildFileObj = {
-      project: {
-        $: {
-          name: '',
-          default: ''
-        }
-      }
-    }
-
-    let currentFile = {
-      name: '',
-      path: ''
-    }
-    let currentTarget = {}
-    let mainTargetsMode = false
-    let otherTargetsMode = false
-
-    const lineData = data.replace(/\r/g, '')
-    const lines = lineData.split('\n')
-
-    for (const line of lines) {
-      console.log(line)
-      if (line.match(/ant\.file\.(?!type)/)) {
-        let lineParts = line.split('ant.file.')
-        if (lineParts.length > 1) {
-          lineParts = lineParts[1].split(' -> ')
-          if (lineParts.length > 1) {
-            currentFile.name = lineParts[0]
-            currentFile.path = lineParts[1]
-          }
-        }
-      }
-
-      if (line.match(/\+Target/)) {
-        const lineParts = line.split('+Target: ')
-        if (lineParts.length > 1) {
-          const targetName = lineParts[1]
-          if (targetName !== '' && !targetName.startsWith(currentFile.name)) {
-            buildFileObj.project.$.targets.push({ name: targetName, sourceFile: currentFile.path })
-          }
-        }
-      }
-
-      if (line.match(/ant.project.name/)) {
-        const lineParts = line.split(' -> ')
-        if (lineParts.length > 1) {
-          buildFileObj.project.$.name = lineParts[1]
-        }
-      }
-
-      if (line.match(/ant.project.default-target/)) {
-        const lineParts = line.split(' -> ')
-        if (lineParts.length > 1) {
-          buildFileObj.project.$.default = lineParts[1]
-        }
-      }
-
-      if (line.match(/^Main targets:/)) {
-        mainTargetsMode = true
-        continue
-      }
-      if (line.match(/^Other targets:/)) {
-        mainTargetsMode = false
-        otherTargetsMode = true
-        continue
-      }
-
-      if (mainTargetsMode || otherTargetsMode) {
-        if (line.match(/^ {3}depends on:/)) {
-          const lineParts = line.split(' depends on: ')
-          if (lineParts.length > 1) {
-            currentTarget.depends = lineParts[1]
-          }
-        }
-
-        // entry is space then target name then two spaces then description
-        for(const target of buildFileObj.project.$.targets) {
-          const regex = new RegExp(`^( ${target.name})`)
-          const match = line.match(regex)
-          // first we must match target name
-          if (match && match.length > 0) {
-            const firstMatch = match[0]
-            const matchLength = firstMatch.length
-
-            // if the line length is equal to matchLength there is no description
-            if (line.length === matchLength) {
-              currentTarget = target
-              continue
-            }
-
-            // the text after the match must start with 2 spaces otherwise we are probably matching the wrong target
-            if (line.length < matchLength + 2 || line[matchLength] !== ' ' || line[matchLength + 1] !== ' ') {
-              continue
-            }
-            currentTarget = target
-            currentTarget.description = line.substring(matchLength + 2).trimStart()
-          }
-        }
-      }
-    }
-
-    return buildFileObj
-  }
-
-  async spawnAnt (buildFileName) {
-    const { spawn } = require('child_process');
-    const child = spawn('S:/Workspaces/vscode-ant/apache-ant/bin/ant.bat', ['-p', '-v', '-d', '-buildfile', buildFileName], { shell: true })
-
-    let data = ""
-    for await (const chunk of child.stdout) {
-        data += chunk
-    }
-    let error = ""
-    for await (const chunk of child.stderr) {
-        error += chunk
-    }
-    const exitCode = await new Promise( (resolve, reject) => {
-        child.on('close', resolve)
-    })
-
-    if( exitCode) {
-        throw new Error( `subprocess error exit ${exitCode}, ${error}`);
-    }
-    return data
-  }
-
-  parseBuildFile (buildFileName) {
+  parseBuildFileDirect (buildFileName) {
     console.debug(`buildFileName= ${buildFileName}`)
     return new Promise((resolve, reject) => {
-      var buildXml = filehelper.getRootFile(this.rootPath, buildFileName)
-      if (filehelper.pathExists(buildXml)) {
+      var buildXml = fileHelper.getRootFile(this.rootPath, buildFileName)
+      if (fileHelper.pathExists(buildXml)) {
         fs.readFile(buildXml, 'utf-8', (err, data) => {
           if (err) {
             return reject(new Error('Error reading build.xml!: ' + err))
