+++ tests/AntWrapper.test.js
@@ -1,0 +1,11 @@
+const context = {
+
+}
+
+const AntWrapper = require("../src/AntWrapper")
+
+describe('AntWrapper', () => {
+  test('should', () => {
+    const antWrapper = new AntWrapper(context)
+  })
+})
